+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+* Flag to ignore upper triangle when calculating distances, instead of always ignore (#20)
+
 ## 1.4.2 - 3 June 2016
 
 ### Changed